dglib_python/pyproject.toml
[project]
name = "dglib-python"
version = "0.1.0"
description = ""
authors = [
    {name = "creation",email = "bsm100100@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (==0.115.12)",
    "uvicorn (==0.34.2)",
    "python-dotenv (==1.1.0)",
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "apscheduler (==3.11.0)",
    "certifi (==2025.4.26)",
    "charset-normalizer (==3.4.2)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "contourpy (==1.3.2)",
    "cycler (==0.12.1)",
    "filelock (==3.18.0)",
    "fonttools (==4.57.0)",
    "fsspec (==2025.3.2)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "jinja2 (==3.1.6)",
    "kiwisolver (==1.4.8)",
    "markupsafe (==3.0.2)",
    "matplotlib (==3.10.1)",
    "mpmath (==1.3.0)",
    "networkx (==3.4.2)",
    "numpy (==2.2.5)",
    "opencv-python (==4.11.0.86)",
    "opencv-python-headless (==4.11.0.86)",
    "packaging (==25.0)",
    "paho-mqtt (==2.1.0)",
    "pandas (==2.2.3)",
    "pillow (==11.2.1)",
    "psutil (==7.0.0)",
    "py-cpuinfo (==9.0.0)",
    "pydantic (==2.11.4)",
    "pydantic-core (==2.33.2)",
    "pyparsing (==3.2.3)",
    "python-dateutil (==2.9.0.post0)",
    "python-multipart (==0.0.20)",
    "pytz (==2025.2)",
    "pyyaml (==6.0.2)",
    "requests (==2.32.3)",
    "scipy (==1.15.2)",
    "seaborn (==0.13.2)",
    "setuptools (==80.1.0)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "starlette (==0.46.2)",
    "sympy (==1.14.0)",
    "torch (==2.7.0)",
    "torchvision (==0.22.0)",
    "tqdm (==4.67.1)",
    "typing-inspection (==0.4.0)",
    "typing-extensions (==4.13.2)",
    "tzdata (==2025.2)",
    "tzlocal (==5.3.1)",
    "ultralytics (==8.3.124)",
    "ultralytics-thop (==2.0.14)",
    "urllib3 (==2.4.0)",
    "watchfiles (==1.0.5)",
    "websockets (==15.0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "dglib_python.book_main:main"
